+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-- `Toggle`: `tooltip` and `tooltipProps` props. ([@qubis741](https://github.com/qubis741)) in [#2355](https://github.com/teamleadercrm/ui/pull/2355))
+- `Toggle`: `labelTooltip` and `labelTooltipProps` props. ([@qubis741](https://github.com/qubis741)) in [#2355](https://github.com/teamleadercrm/ui/pull/2355))
 
 ### Changed